preserve white space in DDM
